+++ CHANGELOG.rst
@@ -5,10 +5,52 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
-[Unreleased]
-************
+0.4.0
+*****
 
-Added
+New functionality
+-----------------
+
+* Extended map_to_seq seasoning
+* Support for YAML timestamp / Python datetime
+* Support for YAML keys with dashes
+
+Fixes
 -----
 
-* Empty Python project directory structure
+* Much improved error messages
+
+0.3.0
+*****
+
+New functionality
+-----------------
+
+* Support for classes that are represented by a string in the YAML file
+* New unified yatiml.Node interface (API change)
+
+Fixes
+-----
+
+* Small improvements to documentation
+* Miscellaneous small fixes
+
+0.2.0
+*****
+
+New functionality
+-----------------
+
+* Support for enumerations
+* Support for user-defined string types
+
+Fixes
+-----
+
+* Various small tooling fixes
+* Some refactoring
+
+0.1.0
+*****
+
+* Initial release with basic functionality